+++ ReleaseNotes.md
@@ -1,6 +1,9 @@
 # [0.1.0]() (release date here)
 
 * Adds `Expand-Template` which allows for the expansion of a powershell string with dynamic parameters.
+* Adds `Expand-TemplateDirectory` which allows for the recursive expansion of a directory and all of its contents.
+* Adds `New-PsstGenerator` which scaffolds the files needed to create a new Psst template generator.
+* Adds `New-PsstGeneratorModule` which scaffolds the files needed to create a new Psst template module.
 * Adds `Get-NamingConventions` which is a utility function that produces name variations from an upper camel case name.
 * Adds `New-AngularComponent` scaffolding function that creates the associated typescript, html, style, and spec files.
 * Adds `New-AngularFeature` scaffolding function that creates the files for a module, bundle, and optional routing module.